requirements.txt
# ...existing requirements...

fastapi==0.68.0
uvicorn==0.34.0
python-multipart== 0.0.20
face-recognition==1.3.0
numpy==1.24.3
python-dotenv==1.1.0
supabase==2.6.0
pydantic==1.10.21
requests== 2.32.3
# ...existing requirements...
